Moved important functionality to player class; separated yahtzee and yahtzee-cli (cli test method)
